pyproject.toml
[project]
name = "bill-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fitz>=0.0.1.dev2",
    "fuzzywuzzy>=0.18.0",
    "google-genai>=1.26.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langgraph>=0.5.4",
    "langsmith>=0.4.8",
    "mistralai>=1.9.2",
    "pandas>=2.0.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "streamlit>=1.47.0",
]
